+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/SpaceEntity.java
@@ -10,7 +10,7 @@ public abstract class SpaceEntity extends Entity {
     /** Shootable strategy */
     private Shootable shootable;
 
-    public SpaceEntity(EntityType type, LinkedList<Entity> subscriber, EntityTurn turn, Shootable shootable, Missile missile, float x, float y, float speed, int durability) {
+    public SpaceEntity(EntityType type, LinkedList<Entity> subscriber, EntityTurn turn, Shootable shootable, float x, float y, float speed, int durability) {
         super(type, subscriber, turn, x, y, speed, durability);
 
         this.shootable = shootable;