+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/missiles/MissileType.java
@@ -1,0 +1,6 @@
+package com.github.mjkuranda.spaceadventure2.entities.missiles;
+
+public enum MissileType {
+    NONE,
+    LASER
+}