+++ src/main/java/com/github/mjkuranda/spaceadventure2/map/GameMap.java
@@ -1,9 +1,7 @@
 package com.github.mjkuranda.spaceadventure2.map;
 
-import com.github.mjkuranda.spaceadventure2.entities.Asteroid;
-import com.github.mjkuranda.spaceadventure2.entities.Entity;
-import com.github.mjkuranda.spaceadventure2.entities.EntityType;
-import com.github.mjkuranda.spaceadventure2.entities.SpaceEntity;
+import com.github.mjkuranda.spaceadventure2.entities.*;
+import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
 
 import java.util.LinkedList;
 import java.util.Random;
@@ -13,11 +11,17 @@ public class GameMap {
     public static final int X_SIZE = 16;
     public static final int Y_SIZE = 16;
 
+    /** Player spaceship */
+    private SpaceShip player;
+
     /** All lines of map */
     private LinkedList<Entity>[] entityLines;
+    private LinkedList<Missile> missileLines;
 
     public GameMap() {
         initLists();
+
+        player = new SpaceShip(missileLines, X_SIZE / 2.0f, Y_SIZE - 1);
     }
 
     public void spawn(EntityType type) {
@@ -51,6 +55,10 @@ public class GameMap {
         }
     }
 
+    public SpaceShip getPlayer() {
+        return player;
+    }
+
     public LinkedList<Entity>[] getSpaceEntityList() {
         return entityLines;
     }
@@ -70,6 +78,8 @@ public class GameMap {
 
     private void initLists() {
         entityLines = new LinkedList[Y_SIZE];
+        missileLines = new LinkedList<>();
+
         initList(entityLines);
     }