+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/GameState.java
@@ -1,7 +1,9 @@
 package com.github.mjkuranda.spaceadventure2.states;
 
 import com.github.mjkuranda.spaceadventure2.entities.Asteroid;
+import com.github.mjkuranda.spaceadventure2.entities.EntityDirection;
 import com.github.mjkuranda.spaceadventure2.entities.EntityType;
+import com.github.mjkuranda.spaceadventure2.entities.SpaceShip;
 import com.github.mjkuranda.spaceadventure2.map.GameMap;
 import com.github.mjkuranda.spaceadventure2.renderers.ArcadeRenderer;
 import com.github.mjkuranda.spaceadventure2.renderers.Renderer;
@@ -47,6 +49,14 @@ public class GameState extends BasicGameState {
 
         if (in.isKeyPressed(Input.KEY_F)) {
             container.setFullscreen(!container.isFullscreen());
+        }
+
+        if (in.isKeyDown(Input.KEY_A)) {
+            map.getPlayer().move(EntityDirection.LEFT);
+        }
+
+        if (in.isKeyDown(Input.KEY_D)) {
+            map.getPlayer().move(EntityDirection.RIGHT);
         }
 
         float prob = new Random().nextFloat();