+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2023-06-23
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): GameMap and Renderer abstract classes.
+- [@mjkuranda](https://github.com/mjkuranda): ArcadeRenderer that renders 2D map.
+- [@mjkuranda](https://github.com/mjkuranda): Basic Player class and handling keys.
 
 ## [0.3.0] - 2023-06-22
 ### Added